Probuje przez crona robic regularne zapisy do bazy danych, narazie nie dziala
