+++ flyway/sql/V1.5__create_temperature_table.sql
@@ -1,4 +1,4 @@
-CREATE TABLE temperature (
+CREATE TABLE temperatures (
   id            BIGINT          AUTO_INCREMENT PRIMARY KEY,
   sensor_id     VARCHAR(100)    NOT NULL,
   value         DECIMAL(5,3)    NOT NULL,
